+++ packages/world/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 2024-06-16
 
 - Wake up player when mining their bed instead of killing them.
+- Fix: Seeds and saplings in circulation were not being tracked correctly.